+++ character/models/character.py
@@ -4,54 +4,14 @@ from django.db import models
 from django.db.models.functions import Upper
 from django.urls import reverse
 
-from character.models.advancement import Advancement
-from character.models.classes import Class, Proficiencies
-from character.models.equipment import Inventory
-from character.models.races import Alignment, Language, Race, Sense, Size
 from utils.dice import Dice
 
-
-class Ability(models.Model):
-    class Name(models.TextChoices):
-        STRENGTH = "STR", "Strength"
-        DEXTERITY = "DEX", "Dexterity"
-        CONSTITUTION = "CON", "Constitution"
-        INTELLIGENCE = "INT", "Intelligence"
-        WISDOM = "WIS", "Wisdom"
-        CHARISMA = "CHA", "Charisma"
-
-    name = models.CharField(max_length=3, primary_key=True, choices=Name)
-
-    class Meta:
-        verbose_name_plural = "abilities"
-
-
-class Skill(models.Model):
-    class Name(models.TextChoices):
-        ATHLETICS = "Athletics"
-        ACROBATICS = "Acrobatics"
-        SLEIGHT_OF_HAND = "Sleight of Hand"
-        STEALTH = "Stealth"
-        ARCANA = "Arcana"
-        HISTORY = "History"
-        INVESTIGATION = "Investigation"
-        NATURE = "Nature"
-        RELIGION = "Religion"
-        ANIMAL_HANDLING = "Animal Handling"
-        INSIGHT = "Insight"
-        MEDICINE = "Medicine"
-        PERCEPTION = "Perception"
-        SURVIVAL = "Survival"
-        DECEPTION = "Deception"
-        INTIMIDATION = "Intimidation"
-        PERFORMANCE = "Performance"
-        PERSUASION = "Persuasion"
-
-    name = models.CharField(max_length=20, primary_key=True, choices=Name)
-    ability = models.ForeignKey(Ability, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return self.name
+from .abilities import Ability
+from .advancement import Advancement
+from .classes import Class, Proficiencies
+from .equipment import Inventory
+from .races import Alignment, Language, Race, Sense, Size
+from .skills import Skill
 
 
 class Character(models.Model):
@@ -72,18 +32,7 @@ class Character(models.Model):
     max_hp = models.SmallIntegerField(default=100)
     proficiency_bonus = models.SmallIntegerField(default=2)
     skills = models.ManyToManyField(Skill)
-    strength = models.SmallIntegerField(default=0)
-    dexterity = models.SmallIntegerField(default=0)
-    constitution = models.SmallIntegerField(default=0)
-    intelligence = models.SmallIntegerField(default=0)
-    wisdom = models.SmallIntegerField(default=0)
-    charisma = models.SmallIntegerField(default=0)
-    strength_modifier = models.SmallIntegerField(default=0)
-    dexterity_modifier = models.SmallIntegerField(default=0)
-    constitution_modifier = models.SmallIntegerField(default=0)
-    intelligence_modifier = models.SmallIntegerField(default=0)
-    wisdom_modifier = models.SmallIntegerField(default=0)
-    charisma_modifier = models.SmallIntegerField(default=0)
+    abilities = models.ManyToManyField(Ability)
     gender = models.CharField(max_length=1, choices=Gender.choices, default=Gender.MALE)
     ac = models.SmallIntegerField(default=0)
     adult_age = models.SmallIntegerField(null=True, blank=True)